server/requirements.txt
flask
flask-mysqldb
mysql-connector-python